third_party/blink/web_tests/media/media-fragments/TC0038-expected.txt


Title: TC0038
Fragment: 't=%6Ept:3'
Comment: The media is requested from 3 seconds to the end. %6E is equivalent to n.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
